+++ src/runtime/Classes/CodedInputStream.m
@@ -30,9 +30,9 @@
 
 @implementation PBCodedInputStream
 
-const int32_t DEFAULT_RECURSION_LIMIT = 64;
-const int32_t DEFAULT_SIZE_LIMIT = 64 << 20;  // 64MB
-const int32_t BUFFER_SIZE = 4096;
+const NSInteger DEFAULT_RECURSION_LIMIT = 64;
+const NSInteger DEFAULT_SIZE_LIMIT = 64 << 20;  // 64MB
+const NSInteger BUFFER_SIZE = 4096;
 
 @synthesize buffer;
 @synthesize input;
@@ -93,7 +93,7 @@ const int32_t BUFFER_SIZE = 4096;
  * Protocol message parsers use this to read tags, since a protocol message
  * may legally end wherever a tag occurs, and zero is not a valid tag number.
  */
-- (int32_t) readTag {
+- (NSInteger) readTag {
   if (self.isAtEnd) {
     lastTag = 0;
     return 0;
@@ -115,7 +115,7 @@ const int32_t BUFFER_SIZE = 4096;
  * @throws InvalidProtocolBufferException {@code value} does not match the
  *                                        last tag.
  */
-- (void) checkLastTagWas:(int32_t) value {
+- (void) checkLastTagWas:(NSInteger) value {
   if (lastTag != value) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"Invalid End Tag" userInfo:nil];
   }
@@ -127,7 +127,7 @@ const int32_t BUFFER_SIZE = 4096;
  * @return {@code NO} if the tag is an endgroup tag, in which case
  *         nothing is skipped.  Otherwise, returns {@code YES}.
  */
-- (BOOL) skipField:(int32_t) tag {
+- (BOOL) skipField:(NSInteger) tag {
   switch (PBWireFormatGetTagWireType(tag)) {
     case PBWireFormatVarint:
       [self readInt32];
@@ -161,7 +161,7 @@ const int32_t BUFFER_SIZE = 4096;
  */
 - (void) skipMessage {
   while (YES) {
-    int32_t tag = [self readTag];
+    NSInteger tag = [self readTag];
     if (tag == 0 || ![self skipField:tag]) {
       return;
     }
@@ -194,7 +194,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 
 /** Read an {@code int32} field value from the stream. */
-- (int32_t) readInt32 {
+- (NSInteger) readInt32 {
   return [self readRawVarint32];
 }
 
@@ -206,7 +206,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 
 /** Read a {@code fixed32} field value from the stream. */
-- (int32_t) readFixed32 {
+- (NSInteger) readFixed32 {
   return [self readRawLittleEndian32];
 }
 
@@ -219,7 +219,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 /** Read a {@code string} field value from the stream. */
 - (NSString*) readString {
-  int32_t size = [self readRawVarint32];
+  NSInteger size = [self readRawVarint32];
   if (size <= (bufferSize - bufferPos) && size > 0) {
     // Fast path:  We already have the bytes in a contiguous buffer, so
     //   just copy directly from it.
@@ -238,7 +238,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 
 /** Read a {@code group} field value from the stream. */
-- (void)      readGroup:(int32_t) fieldNumber
+- (void)      readGroup:(NSInteger) fieldNumber
                 builder:(id<PBMessage_Builder>) builder
       extensionRegistry:(PBExtensionRegistry*) extensionRegistry {
   if (recursionDepth >= recursionLimit) {
@@ -255,7 +255,7 @@ const int32_t BUFFER_SIZE = 4096;
  * Reads a {@code group} field value from the stream and merges it into the
  * given {@link PBUnknownFieldSet}.
  */
-- (void) readUnknownGroup:(int32_t) fieldNumber
+- (void) readUnknownGroup:(NSInteger) fieldNumber
                   builder:(PBUnknownFieldSet_Builder*) builder {
   if (recursionDepth >= recursionLimit) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"Recursion Limit Exceeded" userInfo:nil];
@@ -270,11 +270,11 @@ const int32_t BUFFER_SIZE = 4096;
 /** Read an embedded message field value from the stream. */
 - (void) readMessage:(id<PBMessage_Builder>) builder
    extensionRegistry:(PBExtensionRegistry*) extensionRegistry {
-  int32_t length = [self readRawVarint32];
+  NSInteger length = [self readRawVarint32];
   if (recursionDepth >= recursionLimit) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"Recursion Limit Exceeded" userInfo:nil];
   }
-  int32_t oldLimit = [self pushLimit:length];
+  NSInteger oldLimit = [self pushLimit:length];
   ++recursionDepth;
   [builder mergeFromCodedInputStream:self extensionRegistry:extensionRegistry];
   [self checkLastTagWas:0];
@@ -285,7 +285,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 /** Read a {@code bytes} field value from the stream. */
 - (NSData*) readData {
-  int32_t size = [self readRawVarint32];
+  NSInteger size = [self readRawVarint32];
   if (size < bufferSize - bufferPos && size > 0) {
     // Fast path:  We already have the bytes in a contiguous buffer, so
     //   just copy directly from it.
@@ -300,7 +300,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 
 /** Read a {@code uint32} field value from the stream. */
-- (int32_t) readUInt32 {
+- (NSInteger) readUInt32 {
   return [self readRawVarint32];
 }
 
@@ -309,13 +309,13 @@ const int32_t BUFFER_SIZE = 4096;
  * Read an enum field value from the stream.  Caller is responsible
  * for converting the numeric value to an actual enum.
  */
-- (int32_t) readEnum {
+- (NSInteger) readEnum {
   return [self readRawVarint32];
 }
 
 
 /** Read an {@code sfixed32} field value from the stream. */
-- (int32_t) readSFixed32 {
+- (NSInteger) readSFixed32 {
   return [self readRawLittleEndian32];
 }
 
@@ -327,7 +327,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 
 /** Read an {@code sint32} field value from the stream. */
-- (int32_t) readSInt32 {
+- (NSInteger) readSInt32 {
   return decodeZigZag32([self readRawVarint32]);
 }
 
@@ -344,12 +344,12 @@ const int32_t BUFFER_SIZE = 4096;
  * Read a raw Varint from the stream.  If larger than 32 bits, discard the
  * upper bits.
  */
-- (int32_t) readRawVarint32 {
+- (NSInteger) readRawVarint32 {
   int8_t tmp = [self readRawByte];
   if (tmp >= 0) {
     return tmp;
   }
-  int32_t result = tmp & 0x7f;
+  NSInteger result = tmp & 0x7f;
   if ((tmp = [self readRawByte]) >= 0) {
     result |= tmp << 7;
   } else {
@@ -381,7 +381,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 /** Read a raw Varint from the stream. */
 - (int64_t) readRawVarint64 {
-  int32_t shift = 0;
+  NSInteger shift = 0;
   int64_t result = 0;
   while (shift < 64) {
     int8_t b = [self readRawByte];
@@ -396,16 +396,16 @@ const int32_t BUFFER_SIZE = 4096;
 
 
 /** Read a 32-bit little-endian integer from the stream. */
-- (int32_t) readRawLittleEndian32 {
+- (NSInteger) readRawLittleEndian32 {
   int8_t b1 = [self readRawByte];
   int8_t b2 = [self readRawByte];
   int8_t b3 = [self readRawByte];
   int8_t b4 = [self readRawByte];
   return
-  (((int32_t)b1 & 0xff)      ) |
-  (((int32_t)b2 & 0xff) <<  8) |
-  (((int32_t)b3 & 0xff) << 16) |
-  (((int32_t)b4 & 0xff) << 24);
+  (((NSInteger)b1 & 0xff)      ) |
+  (((NSInteger)b2 & 0xff) <<  8) |
+  (((NSInteger)b3 & 0xff) << 16) |
+  (((NSInteger)b4 & 0xff) << 24);
 }
 
 
@@ -438,11 +438,11 @@ const int32_t BUFFER_SIZE = 4096;
  *
  * @return the old limit.
  */
-- (int32_t) setRecursionLimit:(int32_t) limit {
+- (NSInteger) setRecursionLimit:(NSInteger) limit {
   if (limit < 0) {
     @throw [NSException exceptionWithName:@"IllegalArgument" reason:@"Recursion limit cannot be negative" userInfo:nil];
   }
-  int32_t oldLimit = recursionLimit;
+  NSInteger oldLimit = recursionLimit;
   recursionLimit = limit;
   return oldLimit;
 }
@@ -459,11 +459,11 @@ const int32_t BUFFER_SIZE = 4096;
  *
  * @return the old limit.
  */
-- (int32_t) setSizeLimit:(int32_t) limit {
+- (NSInteger) setSizeLimit:(NSInteger) limit {
   if (limit < 0) {
     @throw [NSException exceptionWithName:@"IllegalArgument" reason:@"Size limit cannot be negative:" userInfo:nil];
   }
-  int32_t oldLimit = sizeLimit;
+  NSInteger oldLimit = sizeLimit;
   sizeLimit = limit;
   return oldLimit;
 }
@@ -483,12 +483,12 @@ const int32_t BUFFER_SIZE = 4096;
  *
  * @return the old limit.
  */
-- (int32_t) pushLimit:(int32_t) byteLimit {
+- (NSInteger) pushLimit:(NSInteger) byteLimit {
   if (byteLimit < 0) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"negativeSize" userInfo:nil];
   }
   byteLimit += totalBytesRetired + bufferPos;
-  int32_t oldLimit = currentLimit;
+  NSInteger oldLimit = currentLimit;
   if (byteLimit > oldLimit) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"truncatedMessage" userInfo:nil];
   }
@@ -502,7 +502,7 @@ const int32_t BUFFER_SIZE = 4096;
 
 - (void) recomputeBufferSizeAfterLimit {
   bufferSize += bufferSizeAfterLimit;
-  int32_t bufferEnd = totalBytesRetired + bufferSize;
+  NSInteger bufferEnd = totalBytesRetired + bufferSize;
   if (bufferEnd > currentLimit) {
     // Limit is in current buffer.
     bufferSizeAfterLimit = bufferEnd - currentLimit;
@@ -518,7 +518,7 @@ const int32_t BUFFER_SIZE = 4096;
  *
  * @param oldLimit The old limit, as returned by {@code pushLimit}.
  */
-- (void) popLimit:(int32_t) oldLimit {
+- (void) popLimit:(NSInteger) oldLimit {
   currentLimit = oldLimit;
   [self recomputeBufferSizeAfterLimit];
 }
@@ -528,12 +528,12 @@ const int32_t BUFFER_SIZE = 4096;
  * Returns the number of bytes to be read before the current limit.
  * If no limit is set, returns -1.
  */
-- (int32_t) bytesUntilLimit {
+- (NSInteger) bytesUntilLimit {
   if (currentLimit == INT_MAX) {
     return -1;
   }
 
-  int32_t currentAbsolutePosition = totalBytesRetired + bufferPos;
+  NSInteger currentAbsolutePosition = totalBytesRetired + bufferPos;
   return currentLimit - currentAbsolutePosition;
 }
 
@@ -587,7 +587,7 @@ const int32_t BUFFER_SIZE = 4096;
     }
   } else {
     [self recomputeBufferSizeAfterLimit];
-    int32_t totalBytesRead = totalBytesRetired + bufferSize + bufferSizeAfterLimit;
+    NSInteger totalBytesRead = totalBytesRetired + bufferSize + bufferSizeAfterLimit;
     if (totalBytesRead > sizeLimit || totalBytesRead < 0) {
       @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"sizeLimitExceeded" userInfo:nil];
     }
@@ -617,7 +617,7 @@ const int32_t BUFFER_SIZE = 4096;
  * @throws InvalidProtocolBufferException The end of the stream or the current
  *                                        limit was reached.
  */
-- (NSData*) readRawData:(int32_t) size {
+- (NSData*) readRawData:(NSInteger) size {
   if (size < 0) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"negativeSize" userInfo:nil];
   }
@@ -640,7 +640,7 @@ const int32_t BUFFER_SIZE = 4096;
 
     // First copy what we have.
     NSMutableData* bytes = [NSMutableData dataWithLength:size];
-    int32_t pos = bufferSize - bufferPos;
+    NSInteger pos = bufferSize - bufferPos;
     memcpy(bytes.mutableBytes, ((int8_t*)buffer.bytes) + bufferPos, pos);
     bufferPos = bufferSize;
 
@@ -671,8 +671,8 @@ const int32_t BUFFER_SIZE = 4096;
 
     // Remember the buffer markers since we'll have to copy the bytes out of
     // it later.
-    int32_t originalBufferPos = bufferPos;
-    int32_t originalBufferSize = bufferSize;
+    NSInteger originalBufferPos = bufferPos;
+    NSInteger originalBufferSize = bufferSize;
 
     // Mark the current buffer consumed.
     totalBytesRetired += bufferSize;
@@ -680,15 +680,15 @@ const int32_t BUFFER_SIZE = 4096;
     bufferSize = 0;
 
     // Read all the rest of the bytes we need.
-    int32_t sizeLeft = size - (originalBufferSize - originalBufferPos);
+    NSInteger sizeLeft = size - (originalBufferSize - originalBufferPos);
     NSMutableArray* chunks = [NSMutableArray array];
 
     while (sizeLeft > 0) {
       NSMutableData* chunk = [NSMutableData dataWithLength:MIN(sizeLeft, BUFFER_SIZE)];
 
-      int32_t pos = 0;
+      NSInteger pos = 0;
       while (pos < chunk.length) {
-        int32_t n = 0;
+        NSInteger n = 0;
         if (input != nil) {
           n = [input read:(((uint8_t*) chunk.mutableBytes) + pos) maxLength:chunk.length - pos];
         }
@@ -706,7 +706,7 @@ const int32_t BUFFER_SIZE = 4096;
     NSMutableData* bytes = [NSMutableData dataWithLength:size];
 
     // Start by copying the leftover bytes from this.buffer.
-    int32_t pos = originalBufferSize - originalBufferPos;
+    NSInteger pos = originalBufferSize - originalBufferPos;
     memcpy(bytes.mutableBytes, ((int8_t*)buffer.bytes) + originalBufferPos, pos);
 
     // And now all the chunks.
@@ -727,7 +727,7 @@ const int32_t BUFFER_SIZE = 4096;
  * @throws InvalidProtocolBufferException The end of the stream or the current
  *                                        limit was reached.
  */
-- (void) skipRawData:(int32_t) size {
+- (void) skipRawData:(NSInteger) size {
   if (size < 0) {
     @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"negativeSize" userInfo:nil];
   }
@@ -744,7 +744,7 @@ const int32_t BUFFER_SIZE = 4096;
     bufferPos += size;
   } else {
     // Skipping more bytes than are in the buffer.  First skip what we have.
-    int32_t pos = bufferSize - bufferPos;
+    NSInteger pos = bufferSize - bufferPos;
     totalBytesRetired += pos;
     bufferPos = 0;
     bufferSize = 0;
@@ -752,7 +752,7 @@ const int32_t BUFFER_SIZE = 4096;
     // Then skip directly from the InputStream for the rest.
     while (pos < size) {
       NSMutableData* data = [NSMutableData dataWithLength:(size - pos)];
-      int32_t n = (input == nil) ? -1 : (int32_t)[input read:data.mutableBytes maxLength:(size - pos)];
+      NSInteger n = (input == nil) ? -1 : (NSInteger)[input read:data.mutableBytes maxLength:(size - pos)];
       if (n <= 0) {
         @throw [NSException exceptionWithName:@"InvalidProtocolBuffer" reason:@"truncatedMessage" userInfo:nil];
       }
