+++ src/runtime/Classes/Field.m
@@ -51,8 +51,8 @@ static PBField *sDefaultInstance = nil;
 	return sDefaultInstance;
 }
 
-- (int32_t)getSerializedSize:(int32_t)fieldNumber {
-	int32_t result = 0;
+- (NSInteger)getSerializedSize:(NSInteger)fieldNumber {
+	NSInteger result = 0;
 
 	const int64_t *varintValues = (const int64_t *)_varintArray.data;
 	if (varintValues) {
@@ -62,7 +62,7 @@ static PBField *sDefaultInstance = nil;
 		}
 	}
 
-	const int32_t *fixed32Values = (const int32_t *)_fixed32Array.data;
+	const NSInteger *fixed32Values = (const NSInteger *)_fixed32Array.data;
 	if (fixed32Values) {
 		const NSUInteger count = _fixed32Array.count;
 		for (NSUInteger i = 0; i < count; ++i) {
@@ -89,8 +89,8 @@ static PBField *sDefaultInstance = nil;
 	return result;
 }
 
-- (int32_t)getSerializedSizeAsMessageSetExtension:(int32_t)fieldNumber {
-	int32_t result = 0;
+- (NSInteger)getSerializedSizeAsMessageSetExtension:(NSInteger)fieldNumber {
+	NSInteger result = 0;
 
 	for (NSData *value in _lengthDelimitedArray) {
 		result += computeRawMessageSetExtensionSize(fieldNumber, value);
@@ -99,7 +99,7 @@ static PBField *sDefaultInstance = nil;
 	return result;
 }
 
-- (void)writeTo:(int32_t)fieldNumber output:(PBCodedOutputStream *) output {
+- (void)writeTo:(NSInteger)fieldNumber output:(PBCodedOutputStream *) output {
 	const int64_t *varintValues = (const int64_t *)_varintArray.data;
 	if (varintValues) {
 		const NSUInteger count = _varintArray.count;
@@ -108,7 +108,7 @@ static PBField *sDefaultInstance = nil;
 		}
 	}
 
-	const int32_t *fixed32Values = (const int32_t *)_fixed32Array.data;
+	const NSInteger *fixed32Values = (const NSInteger *)_fixed32Array.data;
 	if (fixed32Values) {
 		const NSUInteger count = _fixed32Array.count;
 		for (NSUInteger i = 0; i < count; ++i) {
@@ -133,29 +133,29 @@ static PBField *sDefaultInstance = nil;
 	}
 }
 
-- (void)writeDescriptionFor:(int32_t) fieldNumber
+- (void)writeDescriptionFor:(NSInteger) fieldNumber
                          to:(NSMutableString*) output
                  withIndent:(NSString*) indent {
   for (NSNumber* value in self.varintArray) {
-    [output appendFormat:@"%@%d: %qi\n", indent, fieldNumber, value.longLongValue];
+    [output appendFormat:@"%@%ld: %qi\n", indent, (long)fieldNumber, value.longLongValue];
   }
   for (NSNumber* value in self.fixed32Array) {
-    [output appendFormat:@"%@%d: %d\n", indent, fieldNumber, value.intValue];
+    [output appendFormat:@"%@%ld: %d\n", indent, (long)fieldNumber, value.intValue];
   }
   for (NSNumber* value in self.fixed64Array) {
-    [output appendFormat:@"%@%d: %qi\n", indent, fieldNumber, value.longLongValue];
+    [output appendFormat:@"%@%ld: %qi\n", indent, (long)fieldNumber, value.longLongValue];
   }
   for (NSData* value in self.lengthDelimitedArray) {
-    [output appendFormat:@"%@%d: %@\n", indent, fieldNumber, value];
+    [output appendFormat:@"%@%ld: %@\n", indent, (long)fieldNumber, value];
   }
   for (PBUnknownFieldSet* value in self.groupArray) {
-    [output appendFormat:@"%@%d: [\n", indent, fieldNumber];
+    [output appendFormat:@"%@%ld: [\n", indent, (long)fieldNumber];
     [value writeDescriptionTo:output withIndent:[NSString stringWithFormat:@"%@  ", indent]];
     [output appendFormat:@"%@]", indent];
   }
 }
 
-- (void)writeAsMessageSetExtensionTo:(int32_t)fieldNumber output:(PBCodedOutputStream *) output {
+- (void)writeAsMessageSetExtensionTo:(NSInteger)fieldNumber output:(PBCodedOutputStream *) output {
 	for (NSData *value in _lengthDelimitedArray) {
 		[output writeRawMessageSetExtension:fieldNumber value:value];
 	}
