+++ src/runtime/Classes/MutableField.h
@@ -27,7 +27,7 @@
 
 - (PBMutableField *)clear;
 - (PBMutableField *)addVarint:(int64_t)value;
-- (PBMutableField *)addFixed32:(int32_t)value;
+- (PBMutableField *)addFixed32:(NSInteger)value;
 - (PBMutableField *)addFixed64:(int64_t)value;
 - (PBMutableField *)addLengthDelimited:(NSData *)value;
 - (PBMutableField *)addGroup:(PBUnknownFieldSet *)value;
